% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{clean_indices}
\alias{clean_indices}
\title{Cleanse the calculated scores}
\usage{
clean_indices(indices, users_completed, id_cols = user_id)
}
\description{
The most important job here is to remove invalid scores for those with scores
from multiple session, and here just keep the scores from the last session.
}
