% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_data.R
\name{correct_data}
\alias{correct_data}
\alias{correct_device}
\alias{correct_game_dur}
\alias{correct_mst}
\alias{correct_rt}
\title{Correct some specific issues in the data}
\usage{
correct_device(data)

correct_game_dur(data)

correct_mst(data)

correct_rt(data, adjust)
}
\description{
Correct some specific issues in the data
}
