% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fit_cfa}
\alias{fit_cfa}
\title{Fit a confirmatory factor analysis model}
\usage{
fit_cfa(
  config,
  data,
  ...,
  col_manifest = manifest,
  col_latent = latent,
  theory = c("fo", "ho", "bf", "of")
)
}
\arguments{
\item{config}{A data frame with columns \code{latent} and \code{manifest} that
specifies the model. The \code{latent} column specifies the latent variables and
the \code{manifest} column specifies the manifest variables. The \code{latent} column
can be omitted if the model is one-factor model.}

\item{data}{A data frame with the manifest variables.}

\item{...}{Other arguments passed to \code{cfa()}.}

\item{col_manifest, col_latent}{The name of the column in \code{config} that
specifies the manifest and latent variables.}

\item{theory}{The theory of the model. See details.}
}
\value{
The same as \code{\link[lavaan:cfa]{lavaan::cfa()}}.
}
\description{
This is just a wrapper of \code{\link[lavaan:cfa]{lavaan::cfa()}} that allows the model to be
specified in a data frame and allows users to choose from four pre-defined
theories of the model.
}
\details{
Specifically, \code{"fo"} is a first-order model, \code{"ho"} is a higher-order model,
\code{"bf"} is a bi-factor model and \code{"of"} is one-factor model. See Brunner et
al. (2012) for detailed discussion of the models and the naming conventions
used here.
}
