% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_data.R
\name{slice_data}
\alias{slice_data}
\alias{slice_data_trials}
\alias{slice_data_duration}
\alias{slice_data_items}
\alias{slice_data_blocks}
\title{Partition data into roughly equal parts}
\usage{
slice_data_trials(data, num_parts)

slice_data_duration(data, num_parts)

slice_data_items(data, num_parts)

slice_data_blocks(data)
}
\description{
Data partitioning is essential for further task duration calibration, but
there are different data format to partition. Here we use different functions
to complete this goal.
}
\section{Functions}{
\itemize{
\item \code{slice_data_trials()}: For trials format: partition into parts with equal
number of trials (roughly one-minute).

\item \code{slice_data_duration()}: For equal duration format: partition into parts with
equal duration (roughly one-minute).

\item \code{slice_data_items()}: For item-based format: Item means one separate
question. This function will do a basic item analysis to reorder the items
based on the item discrimination.

\item \code{slice_data_blocks()}: For data with explicit blocks: Directly partition into
different blocks.s

}}
