% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate Raw Data}
\usage{
validate_data(data_parsed, require_keyboard, list_names)
}
\arguments{
\item{data_parsed}{Data with parsed raw data.}

\item{require_keyboard}{Logical indicating if keyboard response is required.}

\item{list_names}{A \code{\link[=list]{list()}} of possible column names.}
}
\value{
Validated data of class \code{\link[=data.frame]{data.frame()}}.
}
\description{
This function will ensure only valid data is used for further analysis.
However, no data quality check is performed here.
}
\details{
The validation process contains the following steps:
\enumerate{
\item Keep data with the largest major version only: e.g., remove 2.9.0 and
keep 3.0.0 or 3.1.0.
\item Check if keyboard is used for certain tasks which require keyboard
responses.
\item Check if data names were valid, for some error data contains data from
another game caused by the technical issues.
}
}
