% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fit_cfa}
\alias{fit_cfa}
\title{Fit a confirmatory factor analysis model}
\usage{
fit_cfa(
  config,
  data,
  theory,
  ...,
  col_manifest = manifest,
  col_latent = latent
)
}
\arguments{
\item{config}{A data frame with columns \code{latent} and \code{manifest} that
specifies the model. The \code{latent} column specifies the latent variables and
the \code{manifest} column specifies the manifest variables. The \code{latent} column
can be omitted if the model is one-factor model.}

\item{data}{A data frame with the manifest variables.}

\item{theory}{The theory of the model. Should be one of \code{"fo"} (first-order),
\code{"ho"} (higher-order), \code{"bf"} (bi-factor) and \code{"of"} (one-factor). See
Brunner et al. (2012) for detailed discussion of the models and the naming
conventions used here.}

\item{...}{Other arguments passed to \code{cfa()}.}

\item{col_manifest, col_latent}{The name of the column in \code{config} that
specifies the manifest and latent variables.}
}
\value{
The same as \code{\link[lavaan:cfa]{lavaan::cfa()}}.
}
\description{
This is just a wrapper of \code{\link[lavaan:cfa]{lavaan::cfa()}} that allows the model to be
specified in a data frame and allows users to choose from four pre-defined
theories of the model.
}
\references{
Brunner, M., Nagy, G., & Wilhelm, O. (2012). A Tutorial on Hierarchically
Structured Constructs. Journal of Personality, 80(4), 796–846.
https://doi.org/10.1111/j.1467-6494.2011.00749.x
}
